combine ServiceConfigurationLoader, ServiceConfigurationPersister
